+++ packages/atomic/src/components/search/lazy-index.ts
@@ -20,6 +20,8 @@ export default {
     await import(
       './atomic-result-children-template/atomic-result-children-template.js'
     ),
+  'atomic-result-html': async () =>
+    await import('./atomic-result-html/atomic-result-html.js'),
   'atomic-result-list': async () =>
     await import('./atomic-result-list/atomic-result-list.js'),
   'atomic-result-number': async () =>